refactor: flatten sign-in and sign-out features
