+++ src/app/App.spec.tsx
@@ -46,8 +46,8 @@ vi.mock("@/features/card/read", () => ({
   setCardReadLoading: mocks.setCardReadLoading,
   setCardReadReady: mocks.setCardReadReady,
 }));
-vi.mock("@/features/auth/sign-in", () => ({ loginGoogle: vi.fn() }));
-vi.mock("@/features/auth/sign-out", () => ({ signOutCurrentUser: vi.fn() }));
+vi.mock("@/features/sign-in", () => ({ loginGoogle: vi.fn() }));
+vi.mock("@/features/sign-out", () => ({ signOutCurrentUser: vi.fn() }));
 vi.mock("@/pages/card-form", () => ({ CardFormPage: () => null }));
 vi.mock("@/pages/card-list", () => ({ CardListPage: () => null }));
 vi.mock("@/pages/card-view", () => ({ CardViewPage: () => null }));
